requirements.txt
apache-airflow==2.10.5
kafka-python==2.2.7
confluent-kafka